pyproject.toml
[project]
name = "embedsapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "bs4>=0.0.2",
    "flask[async]>=3.1.1",
    "httpx>=0.28.1",
    "networkx>=3.4.2",
    "pillow>=11.2.1",
]
